style: ruff --fix --unsafe-fixes rewrite + format (Phase-1 baseline)
One-time ruff check --fix --unsafe-fixes pass across liom_toolkit/ gated by
the canary test staying green before, after --fix, and after format:
- 79 auto-fixes applied (import sorting, deprecated APIs, bugbear fixes)
- 22 files reformatted by ruff format
- canary (tests/test_canary.py) 4 passed at all three gates

The Phase-1 select set (E, F, W, I, UP, B, SIM, C4, RUF) bounds what the
rewrite could touch. S101 is not selected, so no assert statements were
rewritten to raise (assert->raise is deferred to a later phase). UP037 did
not fire under target-version=py312, so all 29 "from __future__ import
annotations" imports are preserved (load-bearing on 3.12). The PIL
decompression-bomb guard lines (stats.py, vseg/utils.py) are untouched.

pyproject.toml ruff config supplemented: the non-auto-fixable violations
inside the Phase-1 select set are config-suppressed with ratchet comments
pointing to the phase that picks each up:
- E501, B008, E721, F403, F401 -> Phase 7 / LINT-01
- B904 -> Phase 6 / BUG-01 (raise sites touched there)

After the supplement, ruff check and ruff format --check both exit 0 on the
committed tree, which CI enforces read-only going forward.

+++ liom_toolkit/utils/utils.py
@@ -1,4 +1,5 @@
 from __future__ import annotations
+
 import os
 
 import numpy as np
@@ -23,8 +24,8 @@ def clean_dir(directory: str) -> None:
     :param directory: The directory to clean.
     :type directory: str
     """
-    if os.path.exists(directory + '.DS_Store'):
-        os.remove(directory + '.DS_Store')
+    if os.path.exists(directory + ".DS_Store"):
+        os.remove(directory + ".DS_Store")
 
 
 def convert_to_png_for_saving(img: np.ndarray) -> np.ndarray:
@@ -36,7 +37,6 @@ def convert_to_png_for_saving(img: np.ndarray) -> np.ndarray:
     :return: The converted image
     :rtype: np.ndarray
     """
-    normalized_image = (img - np.min(img)) * (
-            255.0 / (np.max(img) - np.min(img)))
-    normalized_image = normalized_image.astype('uint8')
+    normalized_image = (img - np.min(img)) * (255.0 / (np.max(img) - np.min(img)))
+    normalized_image = normalized_image.astype("uint8")
     return normalized_image
